--- client/packages/system/src/Patient/VaccinationCards/DetailView/index.ts
@@ -1,1 +1,0 @@
-export { VaccinationCardDetailView } from './DetailView';
