--- client/packages/system/src/Patient/api/hooks/keys.ts
@@ -1,1 +1,0 @@
-export const VACCINATION_CARD = 'vaccination-card';
