+++ client/packages/system/src/Patient/api/operations.graphql
@@ -230,43 +230,3 @@ query latestPatientEncounter(
     }
   }
 }
-
-fragment VaccinationCardItem on VaccinationCardItemNode {
-  __typename
-  id
-  vaccineCourseId
-  vaccineCourseDoseId
-  vaccinationId
-  label
-  minAgeMonths
-  vaccinationDate
-  suggestedDate
-  given
-}
-
-fragment VaccinationCard on VaccinationCardNode {
-  __typename
-  id
-  patientFirstName
-  patientLastName
-  programName
-  items {
-    ... on VaccinationCardItemNode {
-      ...VaccinationCardItem
-    }
-  }
-}
-
-query vaccinationCard($storeId: String!, $programEnrolmentId: String!) {
-  vaccinationCard(storeId: $storeId, programEnrolmentId: $programEnrolmentId) {
-    ... on VaccinationCardNode {
-      ...VaccinationCard
-    }
-    ... on NodeError {
-      __typename
-      error {
-        description
-      }
-    }
-  }
-}
