+++ client/packages/system/src/Vaccination/Components/index.ts
@@ -1,1 +1,1 @@
-export * from './VaccinationModal';
+export * from './VaccinationCard';
