+++ client/packages/system/src/Vaccination/api/keys.ts
@@ -1,1 +1,2 @@
 export const VACCINATION = 'vaccination';
+export const VACCINATION_CARD = 'vaccination-card';
