+++ client/packages/system/src/Vaccination/api/usePatientVaccineCard.ts
@@ -1,14 +1,14 @@
 import { useQuery } from '@openmsupply-client/common';
-import { usePatientGraphQL } from '../usePatientGraphQL';
+import { useVaccinationsGraphQL } from './useVaccinationsGraphQL';
 import { VACCINATION_CARD } from './keys';
 
 export const usePatientVaccineCard = (programEnrolmentId: string) => {
-  const { patientApi, storeId } = usePatientGraphQL();
+  const { api, storeId } = useVaccinationsGraphQL();
 
   const { data, isLoading } = useQuery({
     queryKey: [VACCINATION_CARD, programEnrolmentId],
     queryFn: async () => {
-      const result = await patientApi.vaccinationCard({
+      const result = await api.vaccinationCard({
         storeId,
         programEnrolmentId,
       });
